# openapi.yaml
openapi: 3.1.0
info:
  title: Pokémon TCG API
  version: "1.0"
  description: REST API to view, sync, and manage Pokémon TCG sets and user card collections.
servers:
  - url: https://pokeapi.huangtechhub.dev
...
security:
  - oauth2: [openid, profile, email]
  - localAuth: []
paths:
...
  /auth/local/login:
    post:
      summary: Login with password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: JWT token response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: # Changed from 'token' to 'accessToken'
                    type: string
                  email: # Added email to response schema based on image
                    type: string
                  type: # Added type to response schema based on image
                    type: string
                    example: Bearer
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.huangtechhub.dev/application/o/authorize/
          tokenUrl: https://auth.huangtechhub.dev/application/o/token/
          scopes:
            openid: Required for OpenID Connect.
            profile: Access to profile.
            email: Access to email.
    localAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...